Work on cabinetmed module

+++ htdocs/cabinetmed/consultations.php
@@ -20,7 +20,7 @@
  *   \file       htdocs/cabinetmed/consultations.php
  *   \brief      Tab for consultations
  *   \ingroup    cabinetmed
- *   \version    $Id: consultations.php,v 1.24 2011/05/02 19:41:32 eldy Exp $
+ *   \version    $Id: consultations.php,v 1.25 2011/05/02 21:50:01 eldy Exp $
  */
 
 $res=0;
@@ -63,6 +63,8 @@ if (! $sortorder) $sortorder='DESC';
 $limit = $conf->liste_limit;
 
 $consult = new CabinetmedCons($db);
+
+$now=dol_now();
 
 
 /*
@@ -145,7 +147,7 @@ if ($action == 'add' || $action == 'update')
         if ($nbnotempty > 1)
         {
             $error++;
-            $mesgarray[]=$langs->trans("Un seul champ montant possible à la fois");
+            $mesgarray[]=$langs->trans("OnlyOneFieldIsPossible");
         }
         if ($consult->montant_cheque && empty($consult->banque))
         {
@@ -226,9 +228,12 @@ if ($action == 'add' || $action == 'update')
             {
                 $result=$consult->update($user);
 
+                $societe = new Societe($db);
+                $societe->fetch($consult->fk_soc);
+
                 // Search if there is a bank line
                 $bid=0;
-                $sql.= "SELECT b.rowid FROM ".MAIN_DB_PREFIX."bank_url as bu, ".MAIN_DB_PREFIX."bank as b";
+                $sql.= "SELECT b.rowid, b.rappro FROM ".MAIN_DB_PREFIX."bank_url as bu, ".MAIN_DB_PREFIX."bank as b";
                 $sql.= " WHERE bu.url_id = ".$consult->id." AND type = 'consultation'";
                 $sql.= " AND bu.fk_bank = b.rowid";
                 dol_syslog($sql);
@@ -239,6 +244,7 @@ if ($action == 'add' || $action == 'update')
                     if ($obj)
                     {
                         $bid=$obj->rowid;
+                        $rappro=$obj->rappro;
                     }
                 }
                 else
@@ -250,10 +256,12 @@ if ($action == 'add' || $action == 'update')
                 if (! $error)
                 {
                     // If bid
-                    if ($bid)
+                    if ($bid && ! $rappro)
                     {
                         $bankaccountline=new AccountLine($db);
                         $result=$bankaccountline->fetch($bid);
+                        $bank_chq=$bankaccountline->bank_chq;
+                        $fk_bordereau=$bankaccountline->fk_bordereau;
                         $bankaccountline->delete($user);
                     }
 
@@ -278,7 +286,7 @@ if ($action == 'add' || $action == 'update')
                         $bankaccount=new Account($db);
                         $result=$bankaccount->fetch($banque);
                     	if ($result < 0) dol_print_error($db,$bankaccount->error);
-                        $lineid=$bankaccount->addline(dol_now(), $type, $langs->trans("CustomerInvoicePayment"), $amount, $consult->num_cheque, '', $user, $societe->nom, $consult->banque);
+                        $lineid=$bankaccount->addline($now, $type, $langs->trans("CustomerInvoicePayment"), $amount, $consult->num_cheque, '', $user, $societe->nom, $consult->banque);
                         $result1=$bankaccount->add_url_line($lineid,$consult->id,dol_buildpath('/cabinetmed/consultations.php',1).'?action=edit&socid='.$consult->fk_soc.'&id=','Consultation','consultation');
                         $result2=$bankaccount->add_url_line($lineid,$consult->fk_soc,'',$societe->nom,'company');
                         if ($lineid <= 0 || $result1 <= 0 || $result2 <= 0)
@@ -953,5 +961,5 @@ if ($action == '' || $action == 'delete')
 
 $db->close();
 
-llxFooter('$Date: 2011/05/02 19:41:32 $ - $Revision: 1.24 $');
+llxFooter('$Date: 2011/05/02 21:50:01 $ - $Revision: 1.25 $');
 ?>
